More cleanup in Python code.

+++ load.py
@@ -1,7 +1,5 @@
 """Totally definitely EDMCOverlay."""
 
-import os
-import sys
 import time
 import tkinter as tk
 from pathlib import Path
@@ -9,18 +7,15 @@ from subprocess import Popen
 from tkinter import ttk
 
 import myNotebook as nb
-import plug
 from config import config
 from ttkHyperlinkLabel import HyperlinkLabel
 
 import edmcoverlay
 from _logger import logger
 
-_CaptionText: str = "EDMCOverlay for Linux"
-
 logger.debug("Loading plugin...")
 
-
+__CaptionText: str = "EDMCOverlay for Linux"
 __overlay_process: Popen = None
 __xpos_var: tk.IntVar
 __ypos_var: tk.IntVar
@@ -64,40 +59,45 @@ def __start_overlay():
             "edmcintro", "EDMC Overlay for Linux is Ready", "yellow", 30, 165, ttl=10
         )
     else:
-        logger.warning("Overlay is already running, skipping the start.")
+        logger.debug("Overlay is already running, skipping the start.")
 
 
 def __stop_overlay():
     global __overlay_process
     if __overlay_process:
         logger.info("Stopping overlay.")
-        __overlay_process.terminate()
-        __overlay_process.communicate()
+        edmcoverlay.RequestBinaryToStop()
+        time.sleep(1)
+        if __overlay_process.poll() is None:
+            __overlay_process.terminate()
+            __overlay_process.communicate()
         __overlay_process = None
     else:
-        logger.warning("Overlay was not started. Nothing to stop.")
+        logger.debug("Overlay was not started. Nothing to stop.")
 
 
-def plugin_start3(plugin_dir):
-    logger.info("Python code starts.")
-    __start_overlay()
-    return _CaptionText
-
-
+# Reaction to the game start/stop.
 def journal_entry(cmdr, is_beta, system, station, entry, state):
     global __overlay_process
     if entry["event"] in ["LoadGame", "StartUp"] and __overlay_process is None:
-        logger.info("edmcoverlay2: load event received, starting overlay")
+        logger.info("Load event received, ensuring binary overlay is started.")
         __start_overlay()
     elif entry["event"] in ["Shutdown", "ShutDown"]:
-        logger.info("edmcoverlay2: shutdown event received, stopping overlay")
+        logger.info("Shutdown event received, stopping binary overlay.")
         __stop_overlay()
 
 
+# Reaction to EDMC start.
+def plugin_start3(plugin_dir):
+    logger.info("Python code starts.")
+    __start_overlay()
+    return __CaptionText
+
+
+# Reaction to EDMC stop.
 def plugin_stop():
     global __overlay_process
     logger.info("Finishing Python's code.")
-    edmcoverlay.RequestBinaryToStop()
     __stop_overlay()
 
 
@@ -115,7 +115,7 @@ def plugin_prefs(parent: nb.Notebook, cmdr: str, is_beta: bool) -> nb.Frame:
     f0 = nb.Frame(frame)
     HyperlinkLabel(
         f0,
-        text=_CaptionText,
+        text=__CaptionText,
         url="https://github.com/alexzk1/edmcoverlay2",
         background=nb.Label().cget("background"),
         underline=True,
